// src/helpers/enums/message.ts
export enum MessageENUM {
  LENGTH = 'incorrect length in field',
  NOT_URL = 'field bellow require url',
  INVALID_TYPE = ' invalid Type',
  REQUIRED_FIELD = 'A field is required',
  NOT_EXIST = 'A field does not exist',
  EXPIRED = 'A field expired or not exist',
  SYMBOLS = 'field has prohibited symbols',
  FORBIDDEN_VALUE = 'field has forbidden value',
}